fix: resolve all non-archived ty check errors
24 ty errors fixed across 8 files:
- cold_start.py: 12 errors (None guards on ScenarioStats union)
- analysis_cli/cli.py: 3 errors (None guard on variance_decomposition)
- analysis_cli/crossover.py: 3 errors (TypedDict for workload params)
- cli/main.py: removed duplicate _register_analysis with bad import
- dashboard/loader.py: fixed return type list|None -> list
- experiment/calibration.py: fixed return type tuple annotation
- shared/protocols/provisioning.py: fixed get_log return type
- experiment/validation.py: None guard before division
- data/scripts/clf_parser.py: isinstance decode guard

Remaining 48 ty diagnostics are all in archived/ + .venv/ (out of scope)

+++ apps/experiment/experiment/calibration.py
@@ -48,7 +48,7 @@ class CalibrationResult:
 
 def load_test_curl(
     rps: int, duration_sec: int, endpoint: str = "http://localhost:18082/work?duration_ms=10"
-) -> Tuple[int, int, int]:
+) -> Tuple[int, int, float]:
     """
     Run load test using parallel curl requests and measure results via Prometheus.
     Returns: (total_requests, errors, actual_rps)
